ref: Refactor structure
